refactor: 일급컬렉션화 및 책임 분리
